requirements.txt
asgiref==3.6.0
distlib==0.3.6
Django==4.1.6
django-cors-headers==3.14.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
filelock==3.9.0
openpyxl==3.1.2
platformdirs==3.0.0
psycopg2==2.9.7
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7